feat(table improvements): Finalize mobile table design.
