+++ packages/app/src/components/tables/mobile-table.tsx
@@ -1,17 +1,16 @@
 import { AgeGroup } from '~/domain/vaccine/components/age-group';
-import { Anchor, InlineText } from '~/components/typography';
+import { BehaviorAnchor, Cell, HeaderCell, Row, Table } from './components/shared-styled-components';
 import { BehaviorIcon } from '~/domain/behavior/components/behavior-icon';
 import { BehaviorIdentifier } from '~/domain/behavior/logic/behavior-types';
-import { border, BorderProps, compose, display, DisplayProps, minWidth, MinWidthProps, width, WidthProps } from 'styled-system';
 import { Box } from '~/components/base';
-import { colors, NlBehaviorValue, VrBehaviorArchived_20221019Value } from '@corona-dashboard/common';
-import { fontWeights, space } from '~/style/theme';
 import { formatAgeGroupString } from '~/utils/format-age-group-string';
 import { formatBirthyearRangeString } from '~/utils/format-birthyear-range-string';
+import { InlineText } from '~/components/typography';
+import { NlBehaviorValue, VrBehaviorArchived_20221019Value } from '@corona-dashboard/common';
 import { PercentageData, PercentageDataPoint } from './components/percentage-data';
 import { SiteText } from '~/locale';
+import { space } from '~/style/theme';
 import { useIntl } from '~/intl';
-import styled from 'styled-components';
 
 type PercentageData = PercentageDataPoint[][];
 
@@ -38,16 +37,14 @@ export const MobileTable = ({ tableData, headerText, hasAgeGroups, isBehaviourTa
       <Table>
           <thead>
             <Row>
-              <HeaderCell width={{_: '100%', md: 'auto', lg: '300px', xl: '400px'}}>
-                {headerText}
-              </HeaderCell>
+              <HeaderCell width="100%">{headerText}</HeaderCell>
             </Row>
           </thead>
 
           <tbody>
             {tableData.map((item, tableDataIndex) => (
-              <Row key={item.id} display={{ _: 'flex', lg: 'none' }}>
-                <Cell minWidth={{ _: '100%' }}>
+              <Row key={item.id} display="flex">
+                <Cell minWidth="100%">
                   <Box display="flex" margin={`0 ${space[2]} ${space[2]}`}>
                     {isBehaviourTable && (
                       <>
@@ -86,44 +83,3 @@ export const MobileTable = ({ tableData, headerText, hasAgeGroups, isBehaviourTa
     </Box>
   );
 };
-
-const Table = styled.table`
-  border-collapse: collapse;
-  width: 100%;
-`;
-
-type RowProps = DisplayProps;
-
-const Row = styled.tr<RowProps>`
-  flex-wrap: wrap;
-  justify-content: space-between;
-  ${compose(display)};
-`;
-
-type HeaderCellProps = WidthProps & DisplayProps;
-
-const HeaderCell = styled.th<HeaderCellProps>`
-  border-bottom: 1px solid ${colors.gray2};
-  font-weight: ${fontWeights.bold};
-  padding-bottom: ${space[2]};
-  text-align: left;
-  vertical-align: middle;
-  ${compose(width)};
-  ${compose(display)};
-`;
-
-type CellProps = MinWidthProps & BorderProps;
-
-const Cell = styled.td<CellProps>`
-  border-bottom: 1px solid ${colors.gray2};
-  padding: ${space[3]} ${space[0]};
-  vertical-align: middle;
-  ${compose(minWidth)};
-  ${compose(border)};
-`;
-
-const BehaviorAnchor = styled(Anchor)`
-  &:hover {
-    color: ${colors.blue8};
-  }
-`;
